CODE REFACTOR AND V2
I did a whole refactor, trimming off many lines of code, and making life a lot easier.

GONE are the "normalizing" functions. We now have a speed mod array and a 'for' loop. That's all.

The whole reason for this was a 'risky' display. You can now see scroll speed for the next highest speed mod. ezpz